feat: add OfflineAlertBanner
