+++ src/App.tsx
@@ -17,6 +17,7 @@ import { $api } from './api';
 import toTitleCase from './util/string';
 import { useCurrentTime } from './contexts/NowContext';
 import AuthBanner from './components/AuthBanner';
+import AlertBanner from './components/OfflineAlertBanner';
 
 function App() {
     const now = useCurrentTime();
@@ -46,6 +47,7 @@ function App() {
                             CMUEats is now up to date with the official dining website! Sorry for the inconvenience.
                             &gt;_&lt;
                         </div> */}
+                    <AlertBanner />
                     <AuthBanner />
                     <div className="MainContent">
                         <Routes>
